Add inspector, throttled state.
I put state back in with this commit, partly because I'd ironed out the
other bugs. Unfortunately, it still interacts badly with prop function
children. However, I went ahead and shipped it because I've discovered
that you can throttle the state change update on one side, and memoize
the child items on the other, and that keeps the call stack from getting
too big.

This doesn't stop rerenders, unfortunately. You'll want to only grab the
bits of state you want and leave the rest alone in production.